# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: Movie Queue WebApp
  description: Web Frontend for Movie Queue
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: ''
  version: '0.11.5'
paths:
...
  /list/movie_queue:
    get:
      operationId: movie_queue_route
      responses:
        '200':
          description: List Movie Queue Entries
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieQueueRow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    post:
      operationId: movie_queue_update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieQueueUpdateRequest'
        required: true
      responses:
        '201':
          description: Update Movie Queue Entries
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /list/queue/{path}:
    get:
      operationId: movie_queue_show
      responses:
        '200':
          description: Movie Queue
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /list/transcode/queue/{directory}/{file}:
    post:
      operationId: movie_queue_transcode_directory
      responses:
        '201':
          description: Transcode Queue Item
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /list/transcode/queue/{path}:
    post:
      operationId: movie_queue_transcode
      responses:
        '201':
          description: Transcode Queue Item
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
components:
  schemas:
...
    MovieQueueRow:
      type: object
      required:
      - idx
      - collection_idx
      - path
      - show
      properties:
        collection_idx:
          type: string
          format: uuid
        idx:
          type: integer
          format: int32
        last_modified:
          type:
          - string
          - 'null'
          format: date-time
        path:
          $ref: '#/components/schemas/str'
        show:
          $ref: '#/components/schemas/str'
    MovieQueueUpdateRequest:
      type: object
      required:
      - queue
      properties:
        queue:
          type: array
          items:
            $ref: '#/components/schemas/MovieQueueRow'
...
    Pagination:
      type: object
      required:
      - total
      - offset
      - limit
      properties:
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    str:
      type: string